+++ asynchronix/src/model/context.rs
@@ -1,7 +1,10 @@
 use std::fmt;
+use std::time::Duration;
 
 use crate::executor::{Executor, Signal};
-use crate::simulation::{self, LocalScheduler, Mailbox, SchedulerInner};
+use crate::ports::InputFn;
+use crate::simulation::{self, ActionKey, Address, GlobalScheduler, Mailbox, SchedulingError};
+use crate::time::{Deadline, MonotonicTime};
 
 use super::{Model, ProtoModel};
 
@@ -22,7 +25,7 @@ use super::{Model, ProtoModel};
 /// fn self_scheduling_method<'a>(
 ///     &'a mut self,
 ///     arg: MyEventType,
-///     context: &'a Context<Self>
+///     cx: &'a mut Context<Self>
 /// ) -> impl Future<Output=()> + Send + 'a {
 ///     async move {
 ///         /* implementation */
@@ -49,14 +52,14 @@ use super::{Model, ProtoModel};
 ///
 /// impl DelayedGreeter {
 ///     // Triggers a greeting on the output port after some delay [input port].
-///     pub async fn greet_with_delay(&mut self, delay: Duration, context: &Context<Self>) {
-///         let time = context.scheduler.time();
+///     pub async fn greet_with_delay(&mut self, delay: Duration, cx: &mut Context<Self>) {
+///         let time = cx.time();
 ///         let greeting = format!("Hello, this message was scheduled at: {:?}.", time);
 ///
 ///         if delay.is_zero() {
 ///             self.msg_out.send(greeting).await;
 ///         } else {
-///             context.scheduler.schedule_event(delay, Self::send_msg, greeting).unwrap();
+///             cx.schedule_event(delay, Self::send_msg, greeting).unwrap();
 ///         }
 ///     }
 ///
@@ -72,26 +75,293 @@ use super::{Model, ProtoModel};
 // https://github.com/rust-lang/rust/issues/78649
 pub struct Context<M: Model> {
     name: String,
-
-    /// Local scheduler.
-    pub scheduler: LocalScheduler<M>,
+    scheduler: GlobalScheduler,
+    address: Address<M>,
+    origin_id: usize,
 }
 
 impl<M: Model> Context<M> {
     /// Creates a new local context.
-    pub(crate) fn new(name: String, scheduler: LocalScheduler<M>) -> Self {
-        Self { name, scheduler }
+    pub(crate) fn new(name: String, scheduler: GlobalScheduler, address: Address<M>) -> Self {
+        // The only requirement for the origin ID is that it must be (i)
+        // specific to each model and (ii) different from 0 (which is reserved
+        // for the global scheduler). The channel ID of the model mailbox
+        // fulfills this requirement.
+        let origin_id = address.0.channel_id();
+
+        Self {
+            name,
+            scheduler,
+            address,
+            origin_id,
+        }
     }
 
-    /// Returns the model instance name.
+    /// Returns the fully qualified model instance name.
+    ///
+    /// The fully qualified name is made of the unqualified model name, if
+    /// relevant prepended by the dot-separated names of all parent models.
     pub fn name(&self) -> &str {
         &self.name
+    }
+
+    /// Returns the current simulation time.
+    pub fn time(&self) -> MonotonicTime {
+        self.scheduler.time()
+    }
+
+    /// Schedules an event at a future time on this model.
+    ///
+    /// An error is returned if the specified deadline is not in the future of
+    /// the current simulation time.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use std::time::Duration;
+    ///
+    /// use asynchronix::model::{Context, Model};
+    ///
+    /// // A timer.
+    /// pub struct Timer {}
+    ///
+    /// impl Timer {
+    ///     // Sets an alarm [input port].
+    ///     pub fn set(&mut self, setting: Duration, cx: &mut Context<Self>) {
+    ///         if cx.schedule_event(setting, Self::ring, ()).is_err() {
+    ///             println!("The alarm clock can only be set for a future time");
+    ///         }
+    ///     }
+    ///
+    ///     // Rings [private input port].
+    ///     fn ring(&mut self) {
+    ///         println!("Brringggg");
+    ///     }
+    /// }
+    ///
+    /// impl Model for Timer {}
+    /// ```
+    pub fn schedule_event<F, T, S>(
+        &self,
+        deadline: impl Deadline,
+        func: F,
+        arg: T,
+    ) -> Result<(), SchedulingError>
+    where
+        F: for<'a> InputFn<'a, M, T, S>,
+        T: Send + Clone + 'static,
+        S: Send + 'static,
+    {
+        self.scheduler
+            .schedule_event_from(deadline, func, arg, &self.address, self.origin_id)
+    }
+
+    /// Schedules a cancellable event at a future time on this model and returns
+    /// an action key.
+    ///
+    /// An error is returned if the specified deadline is not in the future of
+    /// the current simulation time.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use asynchronix::model::{Context, Model};
+    /// use asynchronix::simulation::ActionKey;
+    /// use asynchronix::time::MonotonicTime;
+    ///
+    /// // An alarm clock that can be cancelled.
+    /// #[derive(Default)]
+    /// pub struct CancellableAlarmClock {
+    ///     event_key: Option<ActionKey>,
+    /// }
+    ///
+    /// impl CancellableAlarmClock {
+    ///     // Sets an alarm [input port].
+    ///     pub fn set(&mut self, setting: MonotonicTime, cx: &mut Context<Self>) {
+    ///         self.cancel();
+    ///         match cx.schedule_keyed_event(setting, Self::ring, ()) {
+    ///             Ok(event_key) => self.event_key = Some(event_key),
+    ///             Err(_) => println!("The alarm clock can only be set for a future time"),
+    ///         };
+    ///     }
+    ///
+    ///     // Cancels the current alarm, if any [input port].
+    ///     pub fn cancel(&mut self) {
+    ///         self.event_key.take().map(|k| k.cancel());
+    ///     }
+    ///
+    ///     // Rings the alarm [private input port].
+    ///     fn ring(&mut self) {
+    ///         println!("Brringggg!");
+    ///     }
+    /// }
+    ///
+    /// impl Model for CancellableAlarmClock {}
+    /// ```
+    pub fn schedule_keyed_event<F, T, S>(
+        &self,
+        deadline: impl Deadline,
+        func: F,
+        arg: T,
+    ) -> Result<ActionKey, SchedulingError>
+    where
+        F: for<'a> InputFn<'a, M, T, S>,
+        T: Send + Clone + 'static,
+        S: Send + 'static,
+    {
+        let event_key = self.scheduler.schedule_keyed_event_from(
+            deadline,
+            func,
+            arg,
+            &self.address,
+            self.origin_id,
+        )?;
+
+        Ok(event_key)
+    }
+
+    /// Schedules a periodically recurring event on this model at a future time.
+    ///
+    /// An error is returned if the specified deadline is not in the future of
+    /// the current simulation time or if the specified period is null.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use std::time::Duration;
+    ///
+    /// use asynchronix::model::{Context, Model};
+    /// use asynchronix::time::MonotonicTime;
+    ///
+    /// // An alarm clock beeping at 1Hz.
+    /// pub struct BeepingAlarmClock {}
+    ///
+    /// impl BeepingAlarmClock {
+    ///     // Sets an alarm [input port].
+    ///     pub fn set(&mut self, setting: MonotonicTime, cx: &mut Context<Self>) {
+    ///         if cx.schedule_periodic_event(
+    ///             setting,
+    ///             Duration::from_secs(1), // 1Hz = 1/1s
+    ///             Self::beep,
+    ///             ()
+    ///         ).is_err() {
+    ///             println!("The alarm clock can only be set for a future time");
+    ///         }
+    ///     }
+    ///
+    ///     // Emits a single beep [private input port].
+    ///     fn beep(&mut self) {
+    ///         println!("Beep!");
+    ///     }
+    /// }
+    ///
+    /// impl Model for BeepingAlarmClock {}
+    /// ```
+    pub fn schedule_periodic_event<F, T, S>(
+        &self,
+        deadline: impl Deadline,
+        period: Duration,
+        func: F,
+        arg: T,
+    ) -> Result<(), SchedulingError>
+    where
+        F: for<'a> InputFn<'a, M, T, S> + Clone,
+        T: Send + Clone + 'static,
+        S: Send + 'static,
+    {
+        self.scheduler.schedule_periodic_event_from(
+            deadline,
+            period,
+            func,
+            arg,
+            &self.address,
+            self.origin_id,
+        )
+    }
+
+    /// Schedules a cancellable, periodically recurring event on this model at a
+    /// future time and returns an action key.
+    ///
+    /// An error is returned if the specified deadline is not in the future of
+    /// the current simulation time or if the specified period is null.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use std::time::Duration;
+    ///
+    /// use asynchronix::model::{Context, Model};
+    /// use asynchronix::simulation::ActionKey;
+    /// use asynchronix::time::MonotonicTime;
+    ///
+    /// // An alarm clock beeping at 1Hz that can be cancelled before it sets off, or
+    /// // stopped after it sets off.
+    /// #[derive(Default)]
+    /// pub struct CancellableBeepingAlarmClock {
+    ///     event_key: Option<ActionKey>,
+    /// }
+    ///
+    /// impl CancellableBeepingAlarmClock {
+    ///     // Sets an alarm [input port].
+    ///     pub fn set(&mut self, setting: MonotonicTime, cx: &mut Context<Self>) {
+    ///         self.cancel();
+    ///         match cx.schedule_keyed_periodic_event(
+    ///             setting,
+    ///             Duration::from_secs(1), // 1Hz = 1/1s
+    ///             Self::beep,
+    ///             ()
+    ///         ) {
+    ///             Ok(event_key) => self.event_key = Some(event_key),
+    ///             Err(_) => println!("The alarm clock can only be set for a future time"),
+    ///         };
+    ///     }
+    ///
+    ///     // Cancels or stops the alarm [input port].
+    ///     pub fn cancel(&mut self) {
+    ///         self.event_key.take().map(|k| k.cancel());
+    ///     }
+    ///
+    ///     // Emits a single beep [private input port].
+    ///     fn beep(&mut self) {
+    ///         println!("Beep!");
+    ///     }
+    /// }
+    ///
+    /// impl Model for CancellableBeepingAlarmClock {}
+    /// ```
+    pub fn schedule_keyed_periodic_event<F, T, S>(
+        &self,
+        deadline: impl Deadline,
+        period: Duration,
+        func: F,
+        arg: T,
+    ) -> Result<ActionKey, SchedulingError>
+    where
+        F: for<'a> InputFn<'a, M, T, S> + Clone,
+        T: Send + Clone + 'static,
+        S: Send + 'static,
+    {
+        let event_key = self.scheduler.schedule_keyed_periodic_event_from(
+            deadline,
+            period,
+            func,
+            arg,
+            &self.address,
+            self.origin_id,
+        )?;
+
+        Ok(event_key)
     }
 }
 
 impl<M: Model> fmt::Debug for Context<M> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("Context").finish_non_exhaustive()
+        f.debug_struct("Context")
+            .field("name", &self.name())
+            .field("time", &self.time())
+            .field("address", &self.address)
+            .field("origin_id", &self.origin_id)
+            .finish_non_exhaustive()
     }
 }
 
@@ -153,7 +423,7 @@ impl<M: Model> fmt::Debug for Context<M> {
 ///
 ///     fn build(
 ///         self,
-///         ctx: &mut BuildContext<Self>)
+///         cx: &mut BuildContext<Self>)
 ///     -> MultiplyBy4 {
 ///         let mut mult = MultiplyBy4 { forward: Output::default() };
 ///         let mut submult1 = MultiplyBy2::default();
@@ -170,8 +440,8 @@ impl<M: Model> fmt::Debug for Context<M> {
 ///         submult1.output.connect(MultiplyBy2::input, &submult2_mbox);
 ///         
 ///         // Add the submodels to the simulation.
-///         ctx.add_submodel(submult1, submult1_mbox, "submultiplier 1");
-///         ctx.add_submodel(submult2, submult2_mbox, "submultiplier 2");
+///         cx.add_submodel(submult1, submult1_mbox, "submultiplier 1");
+///         cx.add_submodel(submult2, submult2_mbox, "submultiplier 2");
 ///
 ///         mult
 ///     }
@@ -180,10 +450,9 @@ impl<M: Model> fmt::Debug for Context<M> {
 /// ```
 #[derive(Debug)]
 pub struct BuildContext<'a, P: ProtoModel> {
-    /// Mailbox of the model.
-    pub mailbox: &'a Mailbox<P::Model>,
+    mailbox: &'a Mailbox<P::Model>,
     name: &'a String,
-    scheduler: &'a SchedulerInner,
+    scheduler: &'a GlobalScheduler,
     executor: &'a Executor,
     abort_signal: &'a Signal,
     model_names: &'a mut Vec<String>,
@@ -194,7 +463,7 @@ impl<'a, P: ProtoModel> BuildContext<'a, P> {
     pub(crate) fn new(
         mailbox: &'a Mailbox<P::Model>,
         name: &'a String,
-        scheduler: &'a SchedulerInner,
+        scheduler: &'a GlobalScheduler,
         executor: &'a Executor,
         abort_signal: &'a Signal,
         model_names: &'a mut Vec<String>,
@@ -215,6 +484,11 @@ impl<'a, P: ProtoModel> BuildContext<'a, P> {
     /// relevant prepended by the dot-separated names of all parent models.
     pub fn name(&self) -> &str {
         self.name
+    }
+
+    /// Returns a handle to the model's mailbox.
+    pub fn address(&self) -> Address<P::Model> {
+        self.mailbox.address()
     }
 
     /// Adds a sub-model to the simulation bench.
