+++ asynchronix/src/ports/input/model_fn.rs
@@ -14,9 +14,9 @@ use super::markers;
 ///
 /// ```ignore
 /// FnOnce(&mut M, T)
-/// FnOnce(&mut M, T, &Context<M>)
+/// FnOnce(&mut M, T, &mut Context<M>)
 /// async fn(&mut M, T)
-/// async fn(&mut M, T, &Context<M>)
+/// async fn(&mut M, T, &mut Context<M>)
 /// where
 ///     M: Model
 /// ```
@@ -34,7 +34,7 @@ pub trait InputFn<'a, M: Model, T, S>: Send + 'static {
     type Future: Future<Output = ()> + Send + 'a;
 
     /// Calls the method.
-    fn call(self, model: &'a mut M, arg: T, context: &'a Context<M>) -> Self::Future;
+    fn call(self, model: &'a mut M, arg: T, cx: &'a mut Context<M>) -> Self::Future;
 }
 
 impl<'a, M, F> InputFn<'a, M, (), markers::WithoutArguments> for F
@@ -44,7 +44,7 @@ where
 {
     type Future = Ready<()>;
 
-    fn call(self, model: &'a mut M, _arg: (), _context: &'a Context<M>) -> Self::Future {
+    fn call(self, model: &'a mut M, _arg: (), _cx: &'a mut Context<M>) -> Self::Future {
         self(model);
 
         ready(())
@@ -58,7 +58,7 @@ where
 {
     type Future = Ready<()>;
 
-    fn call(self, model: &'a mut M, arg: T, _context: &'a Context<M>) -> Self::Future {
+    fn call(self, model: &'a mut M, arg: T, _cx: &'a mut Context<M>) -> Self::Future {
         self(model, arg);
 
         ready(())
@@ -68,12 +68,12 @@ where
 impl<'a, M, T, F> InputFn<'a, M, T, markers::WithContext> for F
 where
     M: Model,
-    F: FnOnce(&'a mut M, T, &'a Context<M>) + Send + 'static,
+    F: FnOnce(&'a mut M, T, &'a mut Context<M>) + Send + 'static,
 {
     type Future = Ready<()>;
 
-    fn call(self, model: &'a mut M, arg: T, context: &'a Context<M>) -> Self::Future {
-        self(model, arg, context);
+    fn call(self, model: &'a mut M, arg: T, cx: &'a mut Context<M>) -> Self::Future {
+        self(model, arg, cx);
 
         ready(())
     }
@@ -87,7 +87,7 @@ where
 {
     type Future = Fut;
 
-    fn call(self, model: &'a mut M, _arg: (), _context: &'a Context<M>) -> Self::Future {
+    fn call(self, model: &'a mut M, _arg: (), _cx: &'a mut Context<M>) -> Self::Future {
         self(model)
     }
 }
@@ -100,7 +100,7 @@ where
 {
     type Future = Fut;
 
-    fn call(self, model: &'a mut M, arg: T, _context: &'a Context<M>) -> Self::Future {
+    fn call(self, model: &'a mut M, arg: T, _cx: &'a mut Context<M>) -> Self::Future {
         self(model, arg)
     }
 }
@@ -109,12 +109,12 @@ impl<'a, M, T, Fut, F> InputFn<'a, M, T, markers::AsyncWithContext> for F
 where
     M: Model,
     Fut: Future<Output = ()> + Send + 'a,
-    F: FnOnce(&'a mut M, T, &'a Context<M>) -> Fut + Send + 'static,
+    F: FnOnce(&'a mut M, T, &'a mut Context<M>) -> Fut + Send + 'static,
 {
     type Future = Fut;
 
-    fn call(self, model: &'a mut M, arg: T, context: &'a Context<M>) -> Self::Future {
-        self(model, arg, context)
+    fn call(self, model: &'a mut M, arg: T, cx: &'a mut Context<M>) -> Self::Future {
+        self(model, arg, cx)
     }
 }
 
@@ -126,7 +126,7 @@ where
 ///
 /// ```ignore
 /// async fn(&mut M, T) -> R
-/// async fn(&mut M, T, &Context<M>) -> R
+/// async fn(&mut M, T, &mut Context<M>) -> R
 /// where
 ///     M: Model
 /// ```
@@ -143,7 +143,7 @@ pub trait ReplierFn<'a, M: Model, T, R, S>: Send + 'static {
     type Future: Future<Output = R> + Send + 'a;
 
     /// Calls the method.
-    fn call(self, model: &'a mut M, arg: T, context: &'a Context<M>) -> Self::Future;
+    fn call(self, model: &'a mut M, arg: T, cx: &'a mut Context<M>) -> Self::Future;
 }
 
 impl<'a, M, R, Fut, F> ReplierFn<'a, M, (), R, markers::AsyncWithoutArguments> for F
@@ -154,7 +154,7 @@ where
 {
     type Future = Fut;
 
-    fn call(self, model: &'a mut M, _arg: (), _context: &'a Context<M>) -> Self::Future {
+    fn call(self, model: &'a mut M, _arg: (), _cx: &'a mut Context<M>) -> Self::Future {
         self(model)
     }
 }
@@ -167,7 +167,7 @@ where
 {
     type Future = Fut;
 
-    fn call(self, model: &'a mut M, arg: T, _context: &'a Context<M>) -> Self::Future {
+    fn call(self, model: &'a mut M, arg: T, _cx: &'a mut Context<M>) -> Self::Future {
         self(model, arg)
     }
 }
@@ -176,11 +176,11 @@ impl<'a, M, T, R, Fut, F> ReplierFn<'a, M, T, R, markers::AsyncWithContext> for 
 where
     M: Model,
     Fut: Future<Output = R> + Send + 'a,
-    F: FnOnce(&'a mut M, T, &'a Context<M>) -> Fut + Send + 'static,
+    F: FnOnce(&'a mut M, T, &'a mut Context<M>) -> Fut + Send + 'static,
 {
     type Future = Fut;
 
-    fn call(self, model: &'a mut M, arg: T, context: &'a Context<M>) -> Self::Future {
-        self(model, arg, context)
+    fn call(self, model: &'a mut M, arg: T, cx: &'a mut Context<M>) -> Self::Future {
+        self(model, arg, cx)
     }
 }
